+++ packages/font/src/pdfkit-standard-fonts.d.ts
@@ -1,0 +1,4 @@
+declare module '@react-pdf/pdfkit/standard-fonts/*' {
+  const fontData: unknown;
+  export default fontData;
+}