+++ packages/layout/src/types/g.ts
@@ -1,4 +1,4 @@
-import { SafeStyle, Style } from '@react-pdf/stylesheet';
+import { SafeStyle, StyleProp } from '@react-pdf/stylesheet';
 import * as P from '@react-pdf/primitives';
 
 import { LineNode, SafeLineNode } from './line';
@@ -17,17 +17,17 @@ import {
 } from './base';
 
 interface GProps extends SVGPresentationAttributes {
-  style?: Style | Style[];
+  style?: StyleProp;
 }
 
 interface SafeGProps extends SafeSVGPresentationAttributes {
-  style?: Style;
+  style?: SafeStyle;
 }
 
 export type GNode = {
   type: typeof P.G;
   props: GProps;
-  style?: Style | Style[];
+  style?: StyleProp;
   box?: never;
   origin?: never;
   yogaNode?: never;